refactor(workspace)!: extract rag-rat-llm — embedder providers, cookbook provisioning, throughput tuning (#706 phase 3)
The SERVING half of index/ai leaves rag-rat-core: the Embedder trait and its
providers (fastembed, model2vec, OpenAI-compatible remote, hash fallback), the
ephemeral cookbook box provisioning contract, and provision-time throughput
tuning. The index-coupled half stays in the engine: reconcile, policy, store,
status, reencode, and the new ai/embedder_select.rs (the glue that resolves
the ACTIVE model/remote config from the index and constructs the right
provider) — that glue reads DB state and scan shapes, so it belongs above the
serving layer, not in it.

Feature plumbing moves with the files: rag-rat-llm owns the fastembed and
model2vec optional deps and features; rag-rat-core's features forward to
them, and the crate's eval feature forwards for the benchmark-only surface
the CLI drives (provision_box_for_benchmark + the measured sweep).

Rides along: the serving-side hash-embedding helpers (hash_embed_text and
support fns) and fastembed_cache_dir move into rag_rat_llm::serving; the
ai-local index_meta kv pair (meta/set_meta) deduplicates into
rag_rat_db::meta; the cookbook test fixtures move with their tests.

Full call-site migration, no re-export shims (core's ai/mod keeps ONLY
pub(crate) internal curation of the llm surface for its reconcile/status
callers); lockstep wiring extended; path-gated workflows include the crate;
first crates.io publish happens locally at the next lockstep release.

No behavior change intended: 2986/2986 tests pass; all four CI clippy gates
green.

BREAKING CHANGE: rag-rat-core no longer exposes the embedder providers,
cookbook provisioning, or benchmark surface on index::ai; library consumers
use the rag-rat-llm crate.

Part of #706.
