+++ .github/workflows/bench-pr-run.yml
@@ -14,6 +14,7 @@ on:
     paths:
       - 'crates/rag-rat-base/**'
       - 'crates/rag-rat-db/**'
+      - 'crates/rag-rat-llm/**'
       - 'crates/rag-rat-papertrail/**'
       - 'crates/rag-rat-core/**'
       - 'Cargo.toml'
