+++ .github/workflows/eval.yml
@@ -14,6 +14,7 @@ on:
     paths:
       - 'crates/rag-rat-base/**'
       - 'crates/rag-rat-db/**'
+      - 'crates/rag-rat-llm/**'
       - 'crates/rag-rat-papertrail/**'
       - 'crates/rag-rat-core/**'
       - 'crates/rag-rat-cli/**'
@@ -27,6 +28,7 @@ on:
     paths:
       - 'crates/rag-rat-base/**'
       - 'crates/rag-rat-db/**'
+      - 'crates/rag-rat-llm/**'
       - 'crates/rag-rat-papertrail/**'
       - 'crates/rag-rat-core/**'
       - 'crates/rag-rat-cli/**'
