+++ crates/rag-rat-core/src/search/semantic.rs
@@ -3,4 +3,4 @@
 //! This module remains the search-facing semantic namespace while the runtime
 //! boundary lives next to model manifests and artifact reconciliation.
 
-pub use crate::index::ai::Embedder;
+pub use rag_rat_llm::providers::Embedder;
