+++ crates/rag-rat-llm/tests/fixtures/cookbook/stub_error_event.sh
