+++ crates/rag-rat-llm/tests/fixtures/cookbook/stub_events.sh
