+++ crates/rag-rat-llm/tests/fixtures/cookbook/stub_fail.sh
