+++ crates/rag-rat-llm/tests/fixtures/cookbook/stub_grandchild_lingers.sh
