+++ crates/rag-rat-llm/tests/fixtures/cookbook/stub_hang_grandchild.sh
