+++ crates/rag-rat-llm/tests/fixtures/cookbook/stub_huge_line.sh
