+++ crates/rag-rat-llm/tests/fixtures/cookbook/stub_ok.sh
