+++ crates/rag-rat-llm/tests/fixtures/cookbook/stub_wrapper_exits_early.sh
